Address review comments

+++ include/oneapi/dpl/experimental/kt/internal/gpu_radix_sort_submitters.h
@@ -25,7 +25,7 @@ namespace oneapi::dpl::experimental::kt::gpu::__impl
 // Please see the comment for __parallel_for_submitter for optional kernel name explanation
 //------------------------------------------------------------------------
 
-template <typename _KeyT, ::std::uint32_t __radix_bits, ::std::uint32_t __stage_count,
+template <typename _KeyT, ::std::uint8_t __radix_bits, ::std::uint32_t __stage_count,
           ::std::uint32_t __hist_work_group_count, ::std::uint32_t __hist_work_group_size, bool __is_ascending,
           typename _KernelName>
 struct __radix_sort_onesweep_histogram_submitter;
